feat: update qt6-webengine to 6.6.1+dfsg-2

+++ src/3rdparty/chromium/chrome/common/logging_chrome.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -18,7 +18,7 @@
 #define IPC_MESSAGE_MACROS_LOG_ENABLED
 #include "content/public/common/content_ipc_logging.h"
 #define IPC_LOG_TABLE_ADD_ENTRY(msg_id, logger) \
-    content::RegisterIPCLogger(msg_id, logger)
+  content::RegisterIPCLogger(msg_id, logger)
 #include "chrome/common/all_messages.h"
 #endif
 
@@ -29,11 +29,10 @@
 #include "chrome/common/logging_chrome.h"
 
 #include <fstream>  // NOLINT
-#include <memory>  // NOLINT
-#include <string>  // NOLINT
+#include <memory>   // NOLINT
+#include <string>   // NOLINT
 
 #include "base/base_switches.h"
-#include "base/bind.h"
 #include "base/command_line.h"
 #include "base/compiler_specific.h"
 #include "base/debug/debugger.h"
@@ -41,6 +40,7 @@
 #include "base/environment.h"
 #include "base/files/file_path.h"
 #include "base/files/file_util.h"
+#include "base/functional/bind.h"
 #include "base/metrics/statistics_recorder.h"
 #include "base/path_service.h"
 #include "base/strings/string_number_conversions.h"
@@ -48,6 +48,7 @@
 #include "base/strings/stringprintf.h"
 #include "base/strings/utf_string_conversions.h"
 #include "build/chromeos_buildflags.h"
+#include "chrome/common/buildflags.h"
 #include "chrome/common/chrome_constants.h"
 #include "chrome/common/chrome_paths.h"
 #include "chrome/common/chrome_switches.h"
@@ -62,6 +63,7 @@
 #if BUILDFLAG(IS_WIN)
 #include <initguid.h>
 #include "base/logging_win.h"
+#include "base/process/process_info.h"
 #include "base/syslog_logging.h"
 #include "chrome/common/win/eventlog_messages.h"
 #include "chrome/install_static/install_details.h"
@@ -85,11 +87,20 @@ bool chrome_logging_failed_ = false;
 // InitChromeLogging() and the beginning of CleanupChromeLogging().
 bool chrome_logging_redirected_ = false;
 
+// The directory on which we do rotation of log files instead of switching
+// with symlink. Because this directory doesn't support symlinks and the logic
+// doesn't work correctly.
+#if BUILDFLAG(IS_CHROMEOS_ASH)
+constexpr char kChronosHomeDir[] = "/home/chronos/user/";
+#endif  // BUILDFLAG(IS_CHROMEOS_ASH)
+
 #if BUILDFLAG(IS_WIN)
 // {7FE69228-633E-4f06-80C1-527FEA23E3A7}
 const GUID kChromeTraceProviderName = {
-    0x7fe69228, 0x633e, 0x4f06,
-        { 0x80, 0xc1, 0x52, 0x7f, 0xea, 0x23, 0xe3, 0xa7 } };
+    0x7fe69228,
+    0x633e,
+    0x4f06,
+    {0x80, 0xc1, 0x52, 0x7f, 0xea, 0x23, 0xe3, 0xa7}};
 #endif
 
 // Assertion handler for logging errors that occur when dialogs are
@@ -112,10 +123,8 @@ void SuppressDialogs() {
       base::BindRepeating(SilentRuntimeAssertHandler));
 
 #if BUILDFLAG(IS_WIN)
-  UINT new_flags = SEM_FAILCRITICALERRORS |
-                   SEM_NOGPFAULTERRORBOX |
-                   SEM_NOOPENFILEERRORBOX;
-
+  UINT new_flags =
+      SEM_FAILCRITICALERRORS | SEM_NOGPFAULTERRORBOX | SEM_NOOPENFILEERRORBOX;
   // Preserve existing error mode, as discussed at http://t/dmea
   UINT existing_flags = SetErrorMode(new_flags);
   SetErrorMode(existing_flags | new_flags);
@@ -129,21 +138,25 @@ void SuppressDialogs() {
 LoggingDestination DetermineLoggingDestination(
     const base::CommandLine& command_line) {
 #if BUILDFLAG(IS_FUCHSIA)
-  // On Fuchsia, the default logging mode is the system log in both debug and
-  // release mode.
+  // Fuchsia provides a system log that can be filtered for logs from specific
+  // components (e.g. Chrome), and which is easier to access than logs in a
+  // file in the component's namespace would be, so always use the system log
+  // by default.
   const LoggingDestination kDefaultLoggingMode = LOG_TO_SYSTEM_DEBUG_LOG;
-#elif NDEBUG
+#elif defined(NDEBUG)
+  // In Release builds, log only to the log file.
   const LoggingDestination kDefaultLoggingMode = LOG_TO_FILE;
 #else
+  // In Debug builds log to all destinations, for ease of discovery.
   const LoggingDestination kDefaultLoggingMode = LOG_TO_ALL;
 #endif  // BUILDFLAG(IS_FUCHSIA)
 
-#ifdef NDEBUG
-  bool enable_logging = false;
-  const char* kInvertLoggingSwitch = switches::kEnableLogging;
-#else
+#if BUILDFLAG(CHROME_ENABLE_LOGGING_BY_DEFAULT)
   bool enable_logging = true;
-  const char* kInvertLoggingSwitch = switches::kDisableLogging;
+  const char* const kInvertLoggingSwitch = switches::kDisableLogging;
+#else
+  bool enable_logging = false;
+  const char* const kInvertLoggingSwitch = switches::kEnableLogging;
 #endif
 
   if (command_line.HasSwitch(kInvertLoggingSwitch))
@@ -160,10 +173,64 @@ LoggingDestination DetermineLoggingDestination(
       return LOG_TO_SYSTEM_DEBUG_LOG | LOG_TO_STDERR;
     } else if (logging_destination != "") {
       PLOG(ERROR) << "Invalid logging destination: " << logging_destination;
+#if BUILDFLAG(IS_WIN)
+    } else if (base::IsCurrentProcessInAppContainer() &&
+               !command_line.HasSwitch(switches::kLogFile)) {
+      // Sandboxed appcontainer processes are unable to resolve the default log
+      // file path without asserting.
+      return kDefaultLoggingMode & ~LOG_TO_FILE;
+#endif
     }
   }
   return kDefaultLoggingMode;
 }
+
+#if BUILDFLAG(IS_CHROMEOS)
+bool RotateLogFile(const base::FilePath& target_path) {
+  DCHECK(!target_path.empty());
+  // If the old log file doesn't exist, do nothing.
+  if (!base::PathExists(target_path)) {
+    return true;
+  }
+
+  // Retrieve the creation time of the old log file.
+  base::File::Info info;
+  {
+    // Opens a file, only if it exists.
+    base::File fp(target_path, base::File::FLAG_OPEN | base::File::FLAG_READ);
+    if (!fp.IsValid() || !fp.GetInfo(&info)) {
+      // On failure, keep using the same file.
+      return false;
+    }
+  }
+
+  // Generate the rotated log path name from the creation time.
+  // (eg. "/home/chrome/user/log/chrome_220102-030405")
+  base::Time timestamp = info.creation_time;
+  base::FilePath rotated_path = GenerateTimestampedName(target_path, timestamp);
+
+  // Rare case: if the target path already exists, generate the alternative by
+  // incrementing the timestamp. This may happen when the Chrome restarts
+  // multiple times in a second.
+  while (base::PathExists(rotated_path)) {
+    timestamp += base::Seconds(1);
+    rotated_path = GenerateTimestampedName(target_path, timestamp);
+  }
+
+  // Rename the old log file: |target_path| => |rotated_path|.
+  // We don't use |base::Move|, since we don't consider the inter-filesystem
+  // move in this logic. The current logic depends on the fact that the ctime
+  // won't be changed after rotation, but ctime may be changed on
+  // inter-filesystem move.
+  if (!base::ReplaceFile(target_path, rotated_path, nullptr)) {
+    PLOG(ERROR) << "Failed to rotate the log files: " << target_path << " => "
+                << rotated_path;
+    return false;
+  }
+
+  return true;
+}
+#endif  // BUILDFLAG(IS_CHROMEOS)
 
 #if BUILDFLAG(IS_CHROMEOS_ASH)
 base::FilePath SetUpSymlinkIfNeeded(const base::FilePath& symlink_path,
@@ -234,47 +301,67 @@ void RemoveSymlinkAndLog(const base::FilePath& link_path,
                          const base::FilePath& target_path) {
   if (::unlink(link_path.value().c_str()) == -1)
     DPLOG(WARNING) << "Unable to unlink symlink " << link_path.value();
-  if (::unlink(target_path.value().c_str()) == -1)
+  if (target_path != link_path && ::unlink(target_path.value().c_str()) == -1)
     DPLOG(WARNING) << "Unable to unlink log file " << target_path.value();
 }
 
 base::FilePath GetSessionLogDir(const base::CommandLine& command_line) {
-  base::FilePath log_dir;
-  std::string log_dir_str;
+  std::string log_dir;
   std::unique_ptr<base::Environment> env(base::Environment::Create());
-  if (env->GetVar(env_vars::kSessionLogDir, &log_dir_str) &&
-      !log_dir_str.empty()) {
-    log_dir = base::FilePath(log_dir_str);
-  } else if (command_line.HasSwitch(ash::switches::kLoginProfile)) {
-    base::PathService::Get(chrome::DIR_USER_DATA, &log_dir);
-    base::FilePath profile_dir;
-    std::string login_profile_value =
-        command_line.GetSwitchValueASCII(ash::switches::kLoginProfile);
-    if (login_profile_value == chrome::kLegacyProfileDir ||
-        login_profile_value == chrome::kTestUserProfileDir) {
-      profile_dir = base::FilePath(login_profile_value);
-    } else {
-      // We could not use g_browser_process > profile_helper() here.
-      std::string profile_dir_str = chrome::kProfileDirPrefix;
-      profile_dir_str.append(login_profile_value);
-      profile_dir = base::FilePath(profile_dir_str);
-    }
-    log_dir = log_dir.Append(profile_dir);
-  }
-  return log_dir;
+  if (!env->GetVar(env_vars::kSessionLogDir, &log_dir))
+    NOTREACHED();
+  return base::FilePath(log_dir);
 }
 
 base::FilePath GetSessionLogFile(const base::CommandLine& command_line) {
   return GetSessionLogDir(command_line)
       .Append(GetLogFileName(command_line).BaseName());
 }
-
 #endif  // BUILDFLAG(IS_CHROMEOS_ASH)
+
+#if BUILDFLAG(IS_CHROMEOS)
+base::FilePath SetUpLogFile(const base::FilePath& target_path, bool new_log) {
+#if BUILDFLAG(IS_CHROMEOS_ASH)
+  const bool supports_symlinks =
+      !(target_path.IsAbsolute() &&
+        base::StartsWith(target_path.value(), kChronosHomeDir));
+
+  // TODO(crbug.com/1326369): Remove the old symlink logic.
+  if (supports_symlinks) {
+    // As for now, we keep the original log rotation logic on the file system
+    // which supports symlinks.
+    return SetUpSymlinkIfNeeded(target_path, new_log);
+  }
+#endif  // BUILDFLAG(IS_CHROMEOS_ASH)
+
+  // Chrome OS doesn't support symlinks on this file system, so that it uses
+  // the rotation logic which doesn't use symlinks.
+  if (!new_log) {
+    // Keep using the same log file without doing anything.
+    return target_path;
+  }
+
+  // For backward compatibility, ignore a ".LATEST" extension the way
+  // |SetUpSymlinkIfNeeded()| does.
+  base::FilePath bare_path = target_path;
+  if (target_path.Extension() == ".LATEST") {
+    bare_path = target_path.ReplaceExtension("");
+  }
+
+  // Try to rotate the log.
+  if (!RotateLogFile(bare_path)) {
+    PLOG(ERROR) << "Failed to rotate the log file: " << bare_path.value()
+                << ". Keeping using the same log file without rotating.";
+  }
+
+  return bare_path;
+}
+#endif  // BUILDFLAG(IS_CHROMEOS)
 
 void InitChromeLogging(const base::CommandLine& command_line,
                        OldFileDeletionState delete_old_log_file) {
-  DCHECK(!chrome_logging_initialized_) <<
-    "Attempted to initialize logging when it was already initialized.";
+  DCHECK(!chrome_logging_initialized_)
+      << "Attempted to initialize logging when it was already initialized.";
   LoggingDestination logging_dest = DetermineLoggingDestination(command_line);
   LogLockingState log_locking_state = LOCK_LOG_FILE;
   base::FilePath log_path;
@@ -294,11 +381,11 @@ void InitChromeLogging(const base::CommandLine& command_line,
     if (command_line.HasSwitch(ash::switches::kGuestSession))
       log_path = GetSessionLogFile(command_line);
 
-    // On ChromeOS we log to the symlink.  We force creation of a new
-    // symlink if we've been asked to delete the old log, since that
+    // Prepares a log file.  We rotate the previous log file and prepare a new
+    // log file if we've been asked to delete the old log, since that
     // indicates the start of a new session.
-    target_path = SetUpSymlinkIfNeeded(
-        log_path, delete_old_log_file == DELETE_OLD_LOG_FILE);
+    target_path =
+        SetUpLogFile(log_path, delete_old_log_file == DELETE_OLD_LOG_FILE);
 
     // Because ChromeOS manages the move to a new session by redirecting
     // the link, it shouldn't remove the old file in the logging code,
@@ -319,7 +406,8 @@ void InitChromeLogging(const base::CommandLine& command_line,
 #if BUILDFLAG(IS_CHROMEOS_ASH)
   if (!success) {
     DPLOG(ERROR) << "Unable to initialize logging to " << log_path.value()
-                << " (which should be a link to " << target_path.value() << ")";
+                 << " (which should be a link to " << target_path.value()
+                 << ")";
     RemoveSymlinkAndLog(log_path, target_path);
     chrome_logging_failed_ = true;
     return;
@@ -393,8 +481,10 @@ void CleanupChromeLogging() {
   if (chrome_logging_failed_)
     return;  // We failed to initiailize logging, no cleanup.
 
-  DCHECK(chrome_logging_initialized_) <<
-    "Attempted to clean up logging when it wasn't initialized.";
+  // Logging was not initialized, no cleanup required. This is happening with
+  // the Chrome early exit error paths (i.e Process Singleton).
+  if (!chrome_logging_initialized_)
+    return;
 
   CloseLogFile();
 
@@ -403,15 +493,38 @@ void CleanupChromeLogging() {
 }
 
 base::FilePath GetLogFileName(const base::CommandLine& command_line) {
+  // Try the command line.
   auto filename = command_line.GetSwitchValueNative(switches::kLogFile);
-  if (!filename.empty())
-    return base::FilePath(filename);
+  // Try the environment.
+  if (filename.empty()) {
+    std::string env_filename;
+    base::Environment::Create()->GetVar(env_vars::kLogFileName, &env_filename);
+#if BUILDFLAG(IS_WIN)
+    filename = base::UTF8ToWide(env_filename);
+#else
+    filename = env_filename;
+#endif  // BUILDFLAG(IS_WIN)
+  }
 
-  std::string env_filename;
-  base::Environment::Create()->GetVar(env_vars::kLogFileName, &env_filename);
-  if (!env_filename.empty())
-    return base::FilePath::FromUTF8Unsafe(env_filename);
+  if (!filename.empty()) {
+    base::FilePath candidate_path(filename);
+#if BUILDFLAG(IS_WIN)
+    // Windows requires an absolute path for the --log-file switch. Windows
+    // cannot log to the current directory as it cds() to the exe's directory
+    // earlier than this function runs.
+    candidate_path = candidate_path.NormalizePathSeparators();
+    if (candidate_path.IsAbsolute()) {
+      return candidate_path;
+    } else {
+      PLOG(ERROR) << "Invalid logging destination: " << filename;
+    }
+#else
+    return candidate_path;
+#endif  // BUILDFLAG(IS_WIN)
+  }
 
+  // If command line and environment do not provide a log file we can use,
+  // fallback to the default.
   const base::FilePath log_filename(FILE_PATH_LITERAL("chrome_debug.log"));
   base::FilePath log_path;
 
@@ -419,8 +532,13 @@ base::FilePath GetLogFileName(const base::CommandLine& command_line) {
     log_path = log_path.Append(log_filename);
     return log_path;
   } else {
-    // error with path service, just use some default file somewhere
+#if BUILDFLAG(IS_WIN)
+    // On Windows we cannot use a non-absolute path so we cannot provide a file.
+    return base::FilePath();
+#else
+    // Error with path service, just use the default in our current directory.
     return log_filename;
+#endif  // BUILDFLAG(IS_WIN)
   }
 }
 
@@ -428,20 +546,17 @@ bool DialogsAreSuppressed() {
   return dialogs_are_suppressed_;
 }
 
-#if BUILDFLAG(IS_CHROMEOS_ASH)
+#if BUILDFLAG(IS_CHROMEOS)
 base::FilePath GenerateTimestampedName(const base::FilePath& base_path,
                                        base::Time timestamp) {
   base::Time::Exploded time_deets;
   timestamp.UTCExplode(&time_deets);
-  std::string suffix = base::StringPrintf("_%02d%02d%02d-%02d%02d%02d",
-                                          time_deets.year,
-                                          time_deets.month,
-                                          time_deets.day_of_month,
-                                          time_deets.hour,
-                                          time_deets.minute,
-                                          time_deets.second);
+  std::string suffix =
+      base::StringPrintf("_%02d%02d%02d-%02d%02d%02d", time_deets.year,
+                         time_deets.month, time_deets.day_of_month,
+                         time_deets.hour, time_deets.minute, time_deets.second);
   return base_path.InsertBeforeExtensionASCII(suffix);
 }
-#endif  // BUILDFLAG(IS_CHROMEOS_ASH)
+#endif  // BUILDFLAG(IS_CHROMEOS)
 
 }  // namespace logging
